+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# v0.0.2 - 2017-10-23
+- Added support for indicating active view
+- Added support for configuring visible views
+
 # v0.0.1 - 2017-07-26
 - Initial release